[refactor] Add StudyPost Controller and Service
