+++ src/main/java/com/example/spot/service/memberstudy/MemberStudyQueryService.java
@@ -33,19 +33,6 @@ public interface MemberStudyQueryService {
     // 금일 회원 출석 여부 불러오기
     StudyQuizResponseDTO.AttendanceListDTO getAllAttendances(Long studyId, Long quizId);
 
-    // 스터디 게시글 목록 불러오기
-    StudyPostResDTO.PostListDTO getAllPosts(PageRequest pageRequest, Long studyId, Theme theme);
-
-    // 스터디 게시글 불러오기
-    StudyPostResDTO.PostDetailDTO getPost(Long studyId, Long postId);
-
-    // 스터디 게시글 댓글 목록 불러오기
-    StudyPostCommentResponseDTO.CommentReplyListDTO getAllComments(Long studyId, Long postId);
-
-
-    // 스터디 이미지 목록 조회
-    StudyImageResponseDTO.ImageListDTO getAllStudyImages(Long studyId);
-
     // 스터디 투표 목록 조회
     StudyVoteResponseDTO.VoteListDTO getAllVotes(Long studyId);
 
@@ -60,4 +47,7 @@ public interface MemberStudyQueryService {
 
     // 스터디 투표 현황 조회
     StudyVoteResponseDTO.CompletedVoteDetailDTO getCompletedVoteDetail(Long studyId, Long voteId);
+
+    // 스터디 이미지 목록 조회
+    StudyImageResponseDTO.ImageListDTO getAllStudyImages(Long studyId);
 }
