+++ src/main/java/com/example/spot/service/memberstudy/MemberStudyQueryServiceImpl.java
@@ -168,114 +168,6 @@ public class MemberStudyQueryServiceImpl implements MemberStudyQueryService {
 
     }
 
-/* ----------------------------- 스터디 게시글 관련 API ------------------------------------- */
-
-    @Override
-    public StudyPostResDTO.PostListDTO getAllPosts(PageRequest pageRequest, Long studyId, Theme theme) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        memberStudyRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, ApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        //=== Feature ===//
-        List<StudyPostResDTO.PostDTO> studyPosts =
-                (theme == null ? studyPostRepository.findAllByStudyId(studyId, pageRequest)
-                        : studyPostRepository.findAllByStudyIdAndTheme(studyId, theme, pageRequest))
-                .stream()
-                .map(StudyPostResDTO.PostDTO::toDTO)
-                .toList();
-
-        return StudyPostResDTO.PostListDTO.toDTO(study, studyPosts);
-
-    }
-
-    @Override
-    @Transactional(readOnly = false)
-    public StudyPostResDTO.PostDetailDTO getPost(Long studyId, Long postId) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        StudyPost studyPost = studyPostRepository.findById(postId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_POST_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        memberStudyRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, ApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        // 해당 스터디의 게시글인지 확인
-        studyPostRepository.findByIdAndStudyId(postId, studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_POST_NOT_FOUND));
-
-        //=== Feature ===//
-        studyPost.plusHitNum();
-        studyPost = studyPostRepository.save(studyPost);
-        memberRepository.save(member);
-        studyRepository.save(study);
-
-        return StudyPostResDTO.PostDetailDTO.toDTO(studyPost);
-    }
-
-    @Override
-    public StudyPostCommentResponseDTO.CommentReplyListDTO getAllComments(Long studyId, Long postId) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        StudyPost studyPost = studyPostRepository.findById(postId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_POST_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        memberStudyRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, ApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        // 해당 스터디의 게시글인지 확인
-        studyPostRepository.findByIdAndStudyId(postId, studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_POST_NOT_FOUND));
-
-        //=== Feature ===//
-
-        return StudyPostCommentResponseDTO.CommentReplyListDTO.toDTO(studyPost.getId(), studyPost.getComments(), defaultImage);
-    }
-
-    @Override
-    public StudyImageResponseDTO.ImageListDTO getAllStudyImages(Long studyId) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._MEMBER_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        memberStudyRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, ApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        //=== Feature ===//
-        return StudyImageResponseDTO.ImageListDTO.toDTO(study);
-    }
-
     /* ----------------------------- 스터디 일정 관련 API ------------------------------------- */
 
     @Override
@@ -504,4 +396,24 @@ public class MemberStudyQueryServiceImpl implements MemberStudyQueryService {
         return StudyVoteResponseDTO.CompletedVoteDetailDTO.toDTO(vote);
     }
 
+/* ----------------------------- 스터디 갤러리 관련 API ------------------------------------- */
+
+    @Override
+    public StudyImageResponseDTO.ImageListDTO getAllStudyImages(Long studyId) {
+
+        //=== Exception ===//
+        Long memberId = SecurityUtils.getCurrentUserId();
+        SecurityUtils.verifyUserId(memberId);
+        Study study = studyRepository.findById(studyId)
+                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new StudyHandler(ErrorStatus._MEMBER_NOT_FOUND));
+
+        // 로그인한 회원이 스터디 회원인지 확인
+        memberStudyRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, ApplicationStatus.APPROVED)
+                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
+
+        //=== Feature ===//
+        return StudyImageResponseDTO.ImageListDTO.toDTO(study);
+    }
 }
