+++ src/main/java/com/example/spot/web/dto/memberstudy/request/StudyVoteRequestDTO.java
@@ -40,7 +40,7 @@ public class StudyVoteRequestDTO {
 
         @TextLength(min = 1, max = 255)
         private String title;
-        
+
         private List<OptionDTO> options;
         private Boolean isMultipleChoice;
         private LocalDateTime finishedAt;
